Move logic of AbstractProjectionItemIterable to a trait to allow easy serialization with JMS serializer
Adjust tests
